requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Pillow==9.3.0
numpy==1.23.5
matplotlib==3.6.2
rembg==2.0.32